+++ src/app/app.module.ts
@@ -7,11 +7,13 @@ import { CartModule } from './cart/cart.module';
 import { ProductsModule } from './product/products.module';
 import { OrderModule } from './order/order.module';
 import { SharedModule } from './shared/shared.module';
+import { AboutComponent } from './layout/components/about/about.component';
 
 @NgModule({
   declarations: [
     AppComponent,
     FirstComponent,
+    AboutComponent
   ],
   imports: [
     BrowserModule,